add: Pages and Layout
* Pages components to files
* Layout components into files
