+++ src/view/Contact.tsx
@@ -1,0 +1,14 @@
+import { Box, Container } from '@mantine/core';
+
+const Contact = () => {
+  return (
+    <Box p='md'>
+      <Container size='xl'>
+        <h1>Contact</h1>
+        <p>This is a simple example of a React component that renders HTML.</p>
+      </Container>
+    </Box>
+  );
+};
+
+export default Contact;
