+++ src/view/Projects.tsx
@@ -1,0 +1,15 @@
+import { Box, Card, } from '@mantine/core';
+import { Link } from 'react-router-dom';
+
+const Projects = () => {
+  return (
+    <Box>
+      <Card>
+        <h2>Project Name</h2>
+        <p>Project Description.</p>
+      </Card>
+    </Box>
+  );
+};
+
+export default Projects;
